float64 police

+++ image_analogy/img_utils.py
@@ -11,7 +11,7 @@ def load_image(image_path):
 
 # util function to open, resize and format pictures into appropriate tensors
 def preprocess_image(x, img_width, img_height):
-    img = imresize(x, (img_height, img_width), interp='bicubic').astype('float64')
+    img = imresize(x, (img_height, img_width), interp='bicubic').astype(np.float32)
     img = vgg16.img_to_vgg(img)
     img = np.expand_dims(img, axis=0)
     return img
